Refactor: #197: comments - 서비스 구현체 분리작업
